// src/interfaces/http/common/error-codes.ts
export enum ErrorCode {
  EMAIL_ALREADY_REGISTERED = 'EMAIL_ALREADY_REGISTERED',
  OTP_NOT_FOUND = 'OTP_NOT_FOUND',
  OTP_EXPIRED = 'OTP_EXPIRED',
  OTP_INVALID = 'OTP_INVALID',
  OTP_LOCKED = 'OTP_LOCKED',
  OTP_RESEND_LIMIT = 'OTP_RESEND_LIMIT',
  OTP_COOLDOWN = 'OTP_COOLDOWN',
  USERNAME_ALREADY_USED = 'USERNAME_ALREADY_USED',

  USER_NOT_FOUND = 'USER_NOT_FOUND',

  REFRESH_TOKEN_MISSING = 'REFRESH_TOKEN_MISSING',
  REFRESH_TOKEN_INVALID = 'REFRESH_TOKEN_INVALID',
  REFRESH_TOKEN_REUSED = 'REFRESH_TOKEN_REUSED',
  REFRESH_TOKEN_EXPIRED = 'REFRESH_TOKEN_EXPIRED',

  ACCESS_TOKEN_EXPIRED = 'ACCESS_TOKEN_EXPIRED',
}